batch_scoring/local/requirements.txt
boto3>=1.28.0
sagemaker>=2.180.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
catboost>=1.2.0
shap>=0.43.0

